Extract async? out of AsyncContext protocol
This change makes so that sieppari does not need to implement the protocol on
every single object that it might possibly process. In the previous version,
not only primitives needed a protocol implementation but also library side
instances like custom records. Not desirable.

The patch extracts async? out of the protocol and uses satisfy? as
implementation.